src/agents/navigator.py
import asyncio
from playwright.async_api import async_playwright
import random

class AlHaqNavigator:
    def __init__(self, user_data_path):
        with open(user_data_path, 'r') as f:
            self.user_data = f.read()

    async def apply_to_linkedin(self, job_url, resume_path, cover_letter_path):
        """
        Automates LinkedIn 'Easy Apply' flow.
        """
        async with async_playwright() as p:
            # Using stealth-like behavior (human-like delays and mouse movements)
            browser = await p.chromium.launch(headless=False) # Headless=False to avoid bot detection
            context = await browser.new_context(
                user_agent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/122.0.0.0 Safari/537.36"
            )
            page = await context.new_page()

            print(f"Navigating to job: {job_url}")
            await page.goto(job_url)
            await asyncio.sleep(random.uniform(2, 5)) # Human-like pause

            # Check if 'Easy Apply' exists
            apply_button = page.locator("button.jobs-apply-button")
            if await apply_button.count() > 0:
                print("Initiating Easy Apply...")
                await apply_button.click()
                await asyncio.sleep(2)

                # Logic to handle the multi-step form would go here:
                # 1. Upload Resume
                # 2. Upload Cover Letter
                # 3. Answer standard questions (Visa, Location, Experience)
                # 4. Submit

                print(f"Successfully uploaded {resume_path}")
            else:
                print("Easy Apply not available for this job. Manual navigation required.")

            await browser.close()

    async def apply_to_bayt(self, job_url, resume_path):
        """
        Automates Bayt.com application flow.
        """
        # Similar logic for Bayt.com (which is critical for Dubai)
        print(f"Bayt Navigator: Preparing application for {job_url}")
        pass

if __name__ == "__main__":
    navigator = AlHaqNavigator('data/master_profile.json')
    # Example execution:
    # asyncio.run(navigator.apply_to_linkedin('https://linkedin.com/jobs/view/123', 'ai-resume/Ali_Resume.pdf', 'ai-cover-letter/Ali_CL.txt'))
